refactor: section 관련 클래스들 패키지로 분리
