+++ src/main/java/subway/line/LineService.java
@@ -1,6 +1,8 @@
 package subway.line;
 
 import org.springframework.stereotype.Service;
+import subway.section.Section;
+import subway.section.SectionService;
 import subway.station.Station;
 import subway.station.StationDao;
