+++ src/main/java/subway/section/SectionService.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.section;
 
 import org.springframework.stereotype.Service;
